requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
distro==1.9.0
docstring_parser==0.17.0
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
packaging==26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.9.2
python-telegram-bot==22.7
replicate==1.0.7
requests==2.32.5
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
